refactor: separate log module
