+++ .size-limit.json
@@ -16,7 +16,7 @@
   {
     "name": "dts libdefs",
     "path": "build/*.d.ts",
-    "limit": "39.1 kB",
+    "limit": "39.15 kB",
     "brotli": false,
     "gzip": false
   },
