+++ test/all.test.js
@@ -20,6 +20,7 @@ import './export.test.js'
 import './global.test.js'
 import './goods.test.js'
 import './index.test.js'
+import './log.test.ts'
 import './md.test.ts'
 import './package.test.js'
 import './util.test.js'
